+++ .gitignore
@@ -29,6 +29,7 @@ out/
 !**/src/main/**/out/
 !**/src/test/**/out/
 
+
 ### NetBeans ###
 /nbproject/private/
 /nbbuild/
@@ -38,3 +39,4 @@ out/
 
 ### VS Code ###
 .vscode/
+**/application.properties
